LI: Use cell->vertex interp operator for upperSurface
This commits makes use of the recently added operator
mpas_cells_to_points_using_baryweights for interpolating
upperSurface to upperSurfaceVertex.  The old method I had
implemented gave garbage values for obtuse triangles, but this method
does not. It requires setting up the interpolation weights on init with
mpas_calculate_barycentric_weights_for_points.

The old method is retained as an alternative (since the new method does
not work across periodic edges).  The method can be controlled with the
config_upperSurfaceVertex_method namelist option.  The new method is the
default.