Create cp protocol for mirror command in setup_testcase
This commit introduces a new protocol for the mirror command that copies
a file from an existing location on the local machine.  It makes use of
subpaths defined by the test case infrastructure.

As part of this change, I also introduced a new function called
expand_path() that converts the path as listed in a .xml file to an
absolute path on the local machine.  This conversion was currently
duplicated in three places and the new cp protocol added a fourth, so
creating a function removes code duplication.