Round adaptive timestep down to nearest second
Fractional seconds can lead to problems because MPAS timekeeping does
not fully support them.  Specifically, timestamps do not include them,
so fractional seconds are lost on a restart, making it not possible for
the restart to be exact if any fields are affected by that level of
precision.  For example, in this branch, daysSinceStart will be
incorrect after a restart if they original run included fractional
seconds, and that will make the Seroussi basal melting parameter to
generate slighly different melt rates since it is a function of
daysSinceStart.

To avoid these complications, this commit forces the adaptive
timestepper to only deal in whole seconds, dropping any fractional
second the adaptive timestep encounters.  (Note it may be preferable to
change the dtSeconds variable from a real to a 8-bit integer.)