Make three temporary fixes for use in the MISMIP+ hydrology experiments.
Changes are made to the thermal and velocity solvers for use in the MISMIP+
with hydrology. These changes are temporary as they are specific to this one
test case. An update should be made for future use in more generalized cases.
First, the basal friction law for heat production is changed to match the sliding
law used, tau_b = beta * basalSpeed**1/3.
Second, comment out the conservation of energy check. This was bringing the model
down so it was eliminated for now.
Third, the betaSolve calculation asks for the minimum between that from the SGH
model and the constant value used in the MISMIP+ experiments, 10002.