Update the MISMIP+ test case with a hydrology option.
Ben Hills' summer project was to experiment with the coupling
between the SGH and the ice dynamics models. This was done
within the framework of MISMIP+. This test case is designed to be
able to reproduce his work.

Along with some minor changes to the
setup files, there are new plotting files and python scripts named
run_ .py which are used to run either the spinup or the experiments
for the chosen parameters.