Changes to support time-averaged output
Code copied from ocean and sea ice cores.
Just monthly and annual time averaging is introduced here.

Note that if a backward_output_offset is used, the backward offset
applied to the start time must result in a year that is greater than or
equal to 0.  In other words, you cannot use a start time of 0000-01-01.

Verified that the time averaged output matches output written at every
timestep and then manually time averaged for both a scalar and map plane
field.  Differences between the two methods are close to roundoff
(1e-13).